+++ CHANGES.md
@@ -58,6 +58,8 @@
   special tags. [2258](https://github.com/jhy/jsoup/issues/2258)
 * An `:nth-child` selector with a negative digit-less step, such as `:nth-child(-n+2)`, would be parsed incorrectly as a
   positive step, and so would not match as expected. [1147](https://github.com/jhy/jsoup/issues/1147)
+* Calling `doc.charset(charset)` on an empty XML document would throw an
+  `IndexOutOfBoundsException`. [2266](https://github.com/jhy/jsoup/issues/2266)
 
 ## 1.18.3 (2024-Dec-02)